backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
curlify==2.2.1
Django==3.2.1
django-cors-headers==3.10.1
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.3.0
drf-yasg==1.20.0
facebook-business==12.0.1
facebookads==2.11.4
frozenlist==1.2.0
gunicorn==19.9.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==5.2.0
openapi-codec==1.3.2
packaging==21.3
psycopg2==2.9.2
pycountry==20.7.3
PyJWT==2.3.0
pyparsing==3.0.6
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.18
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
yarl==1.7.2
